ecom/backend/requirements.txt
django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
django-filter==24.3
requests==2.32.3
Faker==29.0.0
django-admin-extra-buttons==1.5.8
factory-boy